notmuch-view-html: Support email deep in a thread
I suspect that notmuch-show builds the thread structure in the JSON
output an leave the emails that are not matched as null objects. This is
an empirical observation from a failure I just had when trying to view
an e-mail that's a reply in a thread.

I did not try that much to understand the structure of the JSON output
from notmuch-show, but looking into the structure of the example I just
inspected, I seems like this find_email() implementation will suffice
fow now (until another bug is found!).