gitmergetool-nvim-with-diff-base: Use script file
Instead of adding each command as a "+" argument to nvim, let's add all
of them into a script file and then have a single "+" argument to source
it. This makes things more readable and makes us not subject to the
limitations of "+" arguments (e.g. there is a limit on the number of
arguments we can pass).